refactor(165): align android tests to main organization
