+++ app/src/androidTest/java/org/cescfe/numpairs/feature/game/GameRouteTest.kt
@@ -23,7 +23,7 @@ import org.cescfe.numpairs.domain.puzzle.model.Puzzle
 import org.cescfe.numpairs.domain.puzzle.model.Strip
 import org.cescfe.numpairs.domain.puzzle.model.StripItem
 import org.cescfe.numpairs.feature.game.presentation.GameUiState
-import org.cescfe.numpairs.feature.game.ui.GameScreenRobot
+import org.cescfe.numpairs.feature.game.ui.screen.GameScreenRobot
 import org.cescfe.numpairs.ui.theme.NumPairsTheme
 import org.junit.Assert.assertEquals
 import org.junit.Rule
