+++ app/src/androidTest/java/org/cescfe/numpairs/feature/game/ui/screen/GameScreenStripEntryInputTest.kt
@@ -1,4 +1,4 @@
-package org.cescfe.numpairs.feature.game.ui
+package org.cescfe.numpairs.feature.game.ui.screen
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import org.cescfe.numpairs.R
