+++ app/src/androidTest/java/org/cescfe/numpairs/feature/game/ui/screen/GameScreenTestHost.kt
@@ -1,4 +1,4 @@
-package org.cescfe.numpairs.feature.game.ui
+package org.cescfe.numpairs.feature.game.ui.screen
 
 import androidx.activity.ComponentActivity
 import androidx.compose.runtime.collectAsState
@@ -14,7 +14,6 @@ import org.cescfe.numpairs.feature.game.GameHighlightState
 import org.cescfe.numpairs.feature.game.presentation.GameUiState
 import org.cescfe.numpairs.feature.game.presentation.GameViewModel
 import org.cescfe.numpairs.feature.game.presentation.TileOperatorSelectionDialogUiState
-import org.cescfe.numpairs.feature.game.ui.screen.GameScreen
 import org.cescfe.numpairs.ui.theme.NumPairsTheme
 import org.junit.Before
 import org.junit.Rule
